backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
replicate==0.25.1

# Receipt processing dependencies
Pillow==10.0.1
pytesseract==0.3.10
transformers==4.35.0
torch==2.1.0